+++ src/index.ts
@@ -1,3 +1,4 @@
 export * from './constant.js';
 export * from './conversion.js';
+export { parseDuration } from './parseDuration.js';
 export { toParts } from './toParts.js';